fix(main_event): Code Cleanup (#217)
d_rats/ui/event_tab.py:
  New module wtih existing EventTab class.
  Gtk.UIManager does not work with GTK-3

d_rats/mainwindow.py:
  EventTab is now in d_rats.ui.event_tab.

d_rats/ui/main_events.py:
  Remove EventTab class.
